Implemented the image parsing.
Descriptions in the image are parsed as child nodes of the "desc" node.
Note that desc node is the first child of the image node and it must
exist.

The position of each desc node within the image is referenced by the
keyword_to_index_ map defined inside of the image node.