Implemented the faster code formatting by introducing clang-format-server.
Previously, we had to spawn a new clang-format process everytime when we
want to format the code since clang did not provide a library that does
the formatting. Sadly, spawning process is super expensive.

Now, I created the simple clang-format-server which is automatically
spawned upon start. Then, everytime we need to format the code, we can
simply send a request to the format server, using ZMQ.

This change has improved the speed of the overall parsing process
drastically. I also fixed a couple of thread bugs along it.